move tests into __test__ dir
